fix mapping

+++ samples/client/petstore/go/go-petstore-withXml/model_format_test_.go
@@ -28,5 +28,5 @@ type FormatTest struct {
 	DateTime time.Time `json:"dateTime,omitempty" xml:"dateTime"`
 	Uuid string `json:"uuid,omitempty" xml:"uuid"`
 	Password string `json:"password" xml:"password"`
-	BigDecimal BigDecimal `json:"BigDecimal,omitempty" xml:"BigDecimal"`
+	BigDecimal float64 `json:"BigDecimal,omitempty" xml:"BigDecimal"`
 }
