+++ samples/client/petstore/go/go-petstore/model_format_test_.go
@@ -27,5 +27,5 @@ type FormatTest struct {
 	DateTime time.Time `json:"dateTime,omitempty"`
 	Uuid string `json:"uuid,omitempty"`
 	Password string `json:"password"`
-	BigDecimal BigDecimal `json:"BigDecimal,omitempty"`
+	BigDecimal float64 `json:"BigDecimal,omitempty"`
 }
