+++ samples/client/petstore/haskell-http-client/lib/OpenAPIPetstore/ModelLens.hs
@@ -488,7 +488,7 @@ formatTestPasswordL f FormatTest{..} = (\formatTestPassword -> FormatTest { form
 {-# INLINE formatTestPasswordL #-}
 
 -- | 'formatTestBigDecimal' Lens
-formatTestBigDecimalL :: Lens_' FormatTest (Maybe BigDecimal)
+formatTestBigDecimalL :: Lens_' FormatTest (Maybe Double)
 formatTestBigDecimalL f FormatTest{..} = (\formatTestBigDecimal -> FormatTest { formatTestBigDecimal, ..} ) <$> f formatTestBigDecimal
 {-# INLINE formatTestBigDecimalL #-}
